canvas_manage_course/requirements/local.txt
# local environment requirements

-r aws.txt
 
# below are requirements specific to the local environment

ddt==1.1.1
django-debug-toolbar==1.9
django-sslserver==0.21


mock==2.0.0
pyvirtualdisplay==0.2.1
requests-oauthlib==0.8.0
selenium==2.53.2
xlrd==1.0.0

git+ssh://git@github.com/Harvard-University-iCommons/selenium_common.git@v1.4.3#egg=selenium-common==1.4.3